+++ feature/sign-up/src/main/java/com/freetreechair/sign_up/util/BindingAdapters.kt
@@ -1,4 +1,4 @@
-package com.freetreechair.sign_up
+package com.freetreechair.sign_up.util
 
 import android.view.View
 import android.widget.TextView
